refactor: Break up crates/orbit-core/src/command/docs.rs into a command… [ORB-00250]
Break up crates/orbit-core/src/command/docs.rs into a command/docs/ module
