+++ crates/orbit-core/src/command/docs/tests/mod.rs
@@ -1,0 +1,8 @@
+mod add_root;
+mod artifact_ref;
+mod config;
+mod frontmatter;
+mod migrate;
+mod path_util;
+mod search;
+mod walk;
