+++ crates/orbit-core/src/command/docs/tests/search.rs
@@ -1,0 +1,1 @@
+
